move constants into tables directory
