+++ packages/psss/src/constants/index.js
@@ -5,4 +5,3 @@
 
 export * from './colors';
 export * from './images';
-export * from './tables';
