+++ packages/psss/src/containers/Tables/CountrySummaryTable.js
@@ -5,7 +5,7 @@
 
 import React from 'react';
 import { CondensedTableBody, FakeHeader } from '@tupaia/ui-components';
-import { COLUMN_WIDTHS } from '../../constants';
+import { COLUMN_WIDTHS } from './constants';
 import {
   ConnectedTable,
   createTotalCasesAccessor,
