+++ packages/psss/src/containers/Tables/CountryTable.js
@@ -12,7 +12,7 @@ import { Table } from '@tupaia/ui-components';
 import { Alarm, CheckCircleOutline } from '@material-ui/icons';
 import { CountryTableBody } from './CountryTableBody';
 import * as COLORS from '../../constants/colors';
-import { COLUMN_WIDTHS } from '../../constants';
+import { COLUMN_WIDTHS } from './constants';
 import { createTotalCasesAccessor, AlertCell, SitesReportedCell } from '../../components';
 import {
   getCountryWeeks,
