+++ packages/psss/src/containers/Tables/SiteSummaryTable.js
@@ -9,7 +9,7 @@ import { connect } from 'react-redux';
 import Typography from '@material-ui/core/Typography';
 import PropTypes from 'prop-types';
 import { CondensedTableBody, FakeHeader, Table, Button } from '@tupaia/ui-components';
-import { COLUMN_WIDTHS } from '../../constants';
+import { COLUMN_WIDTHS } from './constants';
 import { createTotalCasesAccessor, AlertCell } from '../../components/Table';
 import {
   openWeeklyReportsPanel,
